update boilerplate code (#6)

+++ boilerplate/update.sh
@@ -10,7 +10,9 @@ set -e
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
 OUT="$(mktemp -d)"
-git clone git@github.com:lyft/boilerplate.git "${OUT}"
+trap "rm -fr $OUT" EXIT
+
+git clone git@github.com:flyteorg/boilerplate.git "${OUT}"
 
 echo "Updating the update.sh script."
 cp "${OUT}/boilerplate/update.sh" "${DIR}/update.sh"
@@ -18,7 +20,7 @@ echo ""
 
 
 CONFIG_FILE="${DIR}/update.cfg"
-README="https://github.com/lyft/boilerplate/blob/master/Readme.rst"
+README="https://github.com/flyteorg/boilerplate/blob/master/Readme.rst"
 
 if [ ! -f "$CONFIG_FILE" ]; then
   echo "$CONFIG_FILE not found."
@@ -33,13 +35,34 @@ if [ -z "$REPOSITORY" ]; then
   exit 1
 fi
 
-while read directory; do
-  # TODO: Skip empty lines, whitespace only lines, and comment lines
+while read directory junk; do
+  # Skip comment lines (which can have leading whitespace)
+  if [[ "$directory" == '#'* ]]; then
+    continue
+  fi
+  # Skip blank or whitespace-only lines
+  if [[ "$directory" == "" ]]; then
+    continue
+  fi
+  # Lines like
+  #    valid/path  other_junk
+  # are not acceptable, unless `other_junk` is a comment
+  if [[ "$junk" != "" ]] && [[ "$junk" != '#'* ]]; then
+    echo "Invalid config! Only one directory is allowed per line. Found '$junk'"
+    exit 1
+  fi
+
+  dir_path="${OUT}/boilerplate/${directory}"
+  # Make sure the directory exists
+  if ! [[ -d "$dir_path" ]]; then
+    echo "Invalid boilerplate directory: '$directory'"
+    exit 1
+  fi
+
   echo "***********************************************************************************"
   echo "$directory is configured in update.cfg."
   echo "-----------------------------------------------------------------------------------"
   echo "syncing files from source."
-  dir_path="${OUT}/boilerplate/${directory}"
   rm -rf "${DIR}/${directory}"
   mkdir -p $(dirname "${DIR}/${directory}")
   cp -r "$dir_path" "${DIR}/${directory}"
@@ -50,5 +73,3 @@ while read directory; do
   echo "***********************************************************************************"
   echo ""
 done < "$CONFIG_FILE"
-
-rm -rf "${OUT}"
